revert replicas int32 change
revert the changes brought in
https://github.com/openshift/windows-machine-config-operator/pull/53/files
as the change didn't really fix Bug 1835166

the bug was fixed as a part of upgrading to operator-sdk v0.18

+++ test/e2e/validation_test.go
@@ -284,7 +284,7 @@ func createWindowsMachineConfig(namespace string, isReplicasFieldRequired bool, 
 		},
 	}
 	if isReplicasFieldRequired {
-		wmc.Spec.Replicas = int32(replicasFieldValue)
+		wmc.Spec.Replicas = replicasFieldValue
 	}
 	return wmc
 }
